Change AbstractJoiner and PartitionService to sleep for exponentially increasing time
This introduces ExponentialBackoffSleeper, a utility class that makes it easy to
sleep for exponentially increasing times. This is useful for methods that need to
wait for a particular state to be reached where initially the state should be
checked frequently for responsiveness, but the check interval should be increased
to prevent high CPU usage.